deduplicate some test code about sampling
